move setting components
