+++ src/views/settings/settings.vue
@@ -246,14 +246,14 @@ import { ipcRenderer } from 'electron'
 import { Key } from 'ts-key-enum'
 import { mdiInformationOutline, mdiRefresh, mdiUndoVariant } from '@mdi/js'
 
-import Connection from '@/components/Settings/Connection.vue'
-import Keybinding from '@/components/Settings/Keybinding.vue'
+import LoginHD from '@/common/components/login/hidive.vue'
+import LoginCR from '@/common/components/login/crunchyroll.vue'
 import Checkbox from '@/components/Checkbox.vue'
 import CButton from '@/components/CButton.vue'
 import Icon from '@/components/Icon.vue'
 import Dropdown, { DropdownItem } from '@/components/Form/Dropdown.vue'
-import LoginHD from '@/common/components/login/hidive.vue'
-import LoginCR from '@/common/components/login/crunchyroll.vue'
+import Connection from './components/connection.vue'
+import Keybinding from './components/keybinding.vue'
 
 import { Crunchyroll } from '@/lib/crunchyroll'
 import { getIsUpdateAvailable } from '@/state/app'
